files/dotfiles/.claude/skills/jira/pyproject.toml
[project]
name = "confluence-manager"
version = "0.1.0"
description = "Confluence documentation manager for Claude Code"
requires-python = ">=3.13"
dependencies = [
    "atlassian-python-api>=3.41.0",
    "markdown>=3.4.0",
    "keyring>=24.0.0",
]
